+++ app/src/main/java/com/example/myapplication/PdfModel.java
@@ -1,11 +1,11 @@
 package com.example.myapplication;
 
 public class PdfModel {
-    private int id;
+    private long id;
     private String name;
     private String title;
 
-    public PdfModel(int id, String name, String title) {
+    public PdfModel(long id, String name, String title) {
         this.id = id;
         this.name = name;
         this.title = title;
@@ -19,11 +19,11 @@ public class PdfModel {
         this.title = title;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
