+++ app/src/main/java/com/example/myapplication/VideoModel.java
@@ -1,10 +1,10 @@
 package com.example.myapplication;
 
 public class VideoModel {
-    private int id;
+    private long id;
     private String name, title;
 
-    public VideoModel(int id, String name, String title) {
+    public VideoModel(long id, String name, String title) {
         this.id = id;
         this.name = name;
         this.title = title;
@@ -18,11 +18,11 @@ public class VideoModel {
         this.title = title;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
